requirements.txt
pyyaml
jinja2
minify_html